cache at view scope, drop check for languages setting
(no longer needed with the changes to other code)
