+++ src/main/java/edu/harvard/iq/dataverse/SettingsWrapper.java
@@ -9,19 +9,27 @@ import edu.harvard.iq.dataverse.branding.BrandingUtil;
 import edu.harvard.iq.dataverse.settings.Setting;
 import edu.harvard.iq.dataverse.settings.SettingsServiceBean;
 import edu.harvard.iq.dataverse.settings.SettingsServiceBean.Key;
+import edu.harvard.iq.dataverse.util.BundleUtil;
 import edu.harvard.iq.dataverse.util.MailUtil;
 import edu.harvard.iq.dataverse.util.StringUtil;
 import edu.harvard.iq.dataverse.util.SystemConfig;
 
+import java.io.StringReader;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import javax.ejb.EJB;
 import javax.faces.view.ViewScoped;
 import javax.inject.Named;
+import javax.json.Json;
+import javax.json.JsonArray;
+import javax.json.JsonObject;
+import javax.json.JsonReader;
+import javax.json.JsonValue;
 import javax.mail.internet.InternetAddress;
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -36,7 +44,7 @@ import org.json.JSONObject;
 public class SettingsWrapper implements java.io.Serializable {
 
     @EJB
-    SettingsServiceBean settingService;
+    SettingsServiceBean settingsService;
 
     @EJB
     DataverseServiceBean dataverseService;
@@ -119,7 +127,7 @@ public class SettingsWrapper implements java.io.Serializable {
     private void initSettingsMap() {
         // initialize settings map
         settingsMap = new HashMap<>();
-        for (Setting setting : settingService.listAll()) {
+        for (Setting setting : settingsService.listAll()) {
             settingsMap.put(setting.getName(), setting.getContent());
         }
     }
@@ -290,6 +298,79 @@ public class SettingsWrapper implements java.io.Serializable {
             anonymizedFieldTypes.addAll(Arrays.asList(names.split(",\\s")));
         }
         return anonymizedFieldTypes.contains(df.getDatasetFieldType().getName());
+    }
+
+    Map<String,String> languageMap = null;
+    
+    Map<String, String> getBaseMetadataLanguageMap(boolean refresh) {
+        if (languageMap == null || refresh) {
+            languageMap = new HashMap<String, String>();
+
+            /* If MetadataLanaguages is set, use it.
+             * If not, we can't assume anything and should avoid assuming a metadata language
+             */
+            String mlString = getValueForKey(SettingsServiceBean.Key.MetadataLanguages,"");
+            
+            if(mlString.isEmpty()) {
+                mlString="[]";
+            }
+            JsonReader jsonReader = Json.createReader(new StringReader(mlString));
+            JsonArray languages = jsonReader.readArray();
+            for(JsonValue jv: languages) {
+                JsonObject lang = (JsonObject) jv;
+                languageMap.put(lang.getString("locale"), lang.getString("title"));
+            }
+        }
+        return languageMap;
+    }
+    
+    public Map<String, String> getMetadataLanguages(DvObjectContainer target) {
+        Map<String,String> currentMap = new HashMap<String,String>();
+        currentMap.putAll(getBaseMetadataLanguageMap(true));
+        languageMap.put(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE, getDefaultMetadataLanguageLabel(target));
+        return languageMap;
+    }
+    
+    private String getDefaultMetadataLanguageLabel(DvObjectContainer target) {
+        String mlLabel = Locale.getDefault().getDisplayLanguage();
+        Dataverse parent = target.getOwner();
+        boolean fromAncestor=false;
+        if(parent != null) {
+            mlLabel = parent.getEffectiveMetadataLanguage();
+            //recurse dataverse chain to root and if any have a metadata language set, fromAncestor is true
+            while(parent!=null) {
+                if(!parent.getMetadataLanguage().equals(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE)) {
+                    fromAncestor=true;
+                    break;
+                }
+                parent=parent.getOwner();
+            }
+        }
+        if(mlLabel.equals(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE)) {
+            mlLabel = getBaseMetadataLanguageMap(false).get(getDefaultMetadataLanguage());
+        }
+        if(fromAncestor) {
+            mlLabel = mlLabel + " " + BundleUtil.getStringFromBundle("dataverse.inherited");
+        } else {
+            mlLabel = mlLabel + " " + BundleUtil.getStringFromBundle("dataverse.default");
+        }
+        return mlLabel;
+    }
+    
+    public String getDefaultMetadataLanguage() {
+        Map<String, String> mdMap = getBaseMetadataLanguageMap(false);
+        if(mdMap.size()>=1) {
+            if(mdMap.size()==1) {
+                //One entry - it's the default
+            return (String) mdMap.keySet().toArray()[0];
+            } else {
+                //More than one - :MetadataLanguages is set so we use the default
+                return DvObjectContainer.DEFAULT_METADATA_LANGUAGE_CODE;
+            }
+        } else {
+            // None - :MetadataLanguages is not set so return null to turn off the display (backward compatibility)
+            return null;
+        }
     }
 
 }
