+++ src/main/java/edu/harvard/iq/dataverse/util/SystemConfig.java
@@ -1082,84 +1082,4 @@ public class SystemConfig {
         return System.getProperty("doi.dataciterestapiurlstring", System.getProperty("doi.mdcbaseurlstring", "https://api.datacite.org"));
 	}
 	
-    Map<String,String> languageMap = null;
-    
-    public Map<String, String> getBaseMetadataLanguageMap(boolean refresh) {
-        if (languageMap == null || refresh) {
-            languageMap = new HashMap<String, String>();
-
-            /* If MetadataLanaguages is set, use it.
-             * If not, if there is no Langauages setting, which means the instance is single language, assume that language is also used for metadata
-             * If not and there is a Language setting, we can't assume anything and should avoid assuming a metadata language
-             */
-            String mlString = settingsService.get(SettingsServiceBean.Key.MetadataLanguages.toString(),"");
-            
-            if(mlString.isEmpty()) {
-                if(settingsService.get(SettingsServiceBean.Key.Languages.toString(),"").isEmpty()) {
-                    mlString="[{\"locale\":\"" + BundleUtil.getCurrentLocale().getLanguage()  + "\",\"title\":\""
-                            + BundleUtil.getCurrentLocale().getDisplayLanguage() + "\"}]";
-                } else {
-                    mlString="[]";
-                }
-            }
-            logger.fine("mlString is : " + mlString);
-            JsonReader jsonReader = Json.createReader(new StringReader(mlString));
-            JsonArray languages = jsonReader.readArray();
-            for(JsonValue jv: languages) {
-                JsonObject lang = (JsonObject) jv;
-                languageMap.put(lang.getString("locale"), lang.getString("title"));
-            }
-        }
-        return languageMap;
-    }
-    
-    public Map<String, String> getMetadataLanguages(DvObjectContainer target) {
-        Map<String,String> currentMap = new HashMap<String,String>();
-        currentMap.putAll(getBaseMetadataLanguageMap(true));
-        languageMap.put(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE, getDefaultMetadataLanguageLabel(target));
-        return languageMap;
-    }
-    
-    private String getDefaultMetadataLanguageLabel(DvObjectContainer target) {
-        String mlLabel = Locale.getDefault().getDisplayLanguage();
-        Dataverse parent = target.getOwner();
-        boolean fromAncestor=false;
-        if(parent != null) {
-            mlLabel = parent.getEffectiveMetadataLanguage();
-            //recurse dataverse chain to root and if any have a metadata language set, fromAncestor is true
-            while(parent!=null) {
-                if(!parent.getMetadataLanguage().equals(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE)) {
-                    fromAncestor=true;
-                    break;
-                }
-                parent=parent.getOwner();
-            }
-        }
-        if(mlLabel.equals(DvObjectContainer.UNDEFINED_METADATA_LANGUAGE_CODE)) {
-            mlLabel = getBaseMetadataLanguageMap(false).get(getDefaultMetadataLanguage());
-        }
-        if(fromAncestor) {
-            mlLabel = mlLabel + " " + BundleUtil.getStringFromBundle("dataverse.inherited");
-        } else {
-            mlLabel = mlLabel + " " + BundleUtil.getStringFromBundle("dataverse.default");
-        }
-        return mlLabel;
-    }
-    
-    public String getDefaultMetadataLanguage() {
-        Map<String, String> mdMap = getBaseMetadataLanguageMap(false);
-        if(mdMap.size()>=1) {
-            if(mdMap.size()==1) {
-                //One entry - it's the default
-            return (String) mdMap.keySet().toArray()[0];
-            } else {
-                //More than one - :MetadataLanguages is set so we use the default
-                return DvObjectContainer.DEFAULT_METADATA_LANGUAGE_CODE;
-            }
-        } else {
-            // None - :MetadataLanguages is not set so return null to turn off the display (backward compatibility)
-            return null;
-        }
-    }
-
 }
